Added ResultFileOutputStream class adapter
